requirements.txt
fastapi==0.115.0
uvicorn==0.29.0
streamlit==1.35.0
pdfplumber==0.11.0
PyPDF2==3.0.1
spacy==3.8.3
sentence-transformers==3.0.1
faiss-cpu==1.8.0
pandas==2.2.2
numpy==1.26.4
loguru==0.7.2
pydantic==2.7.1
python-multipart==0.0.9
requests==2.32.3